moved to use deepPartialMatch
