+++ lib/iterator.js
@@ -1,26 +1,7 @@
 'use strict'
 
 var matchingBuckets = require('./matchingBuckets')
-
-function deepPartialMatch (a, b) {
-  var keys
-
-  if (a === b) {
-    return true
-  } else if (typeof a !== 'object' || typeof b !== 'object') {
-    return false
-  }
-
-  keys = Object.keys(a)
-
-  for (var i = 0; i < keys.length; i++) {
-    if (a[keys[i]] && b[keys[i]] && !deepPartialMatch(a[keys[i]], b[keys[i]])) {
-      return false
-    }
-  }
-
-  return true
-}
+var deepPartialMatch = require('./deepPartialMatch')
 
 function Iterator (parent, obj) {
   if (!(this instanceof Iterator)) {
